// src/model/interfaces/cartaPresentacion/cartaPresentacion.ts
export default interface CartaPresentacionDatos {
    cartaId: number
    cartaEstado: number
    cartaFechaRegistro: string
    anioNombre: string
    repDni: string
    repNombreCompleto: string
    repCargoNombre: string
    empresaRuc: number
    empresaNombre: string
    empresaDireccion: string
    empresaDPD: string
    estId: string
    estNombreCompleto: string
    estNivel: string
    estFacultad: string
    estCarrera: string
    estCurso: string
    estSeccion: string
    admNombreCompleto: string
    admCargo: string
}