// src/model/interfaces/convenio/convenioPracticas.ts
export default interface ConvenioPracticas {
    tipoConvenio: number

    nombreEmpresa: string
    rucEmpresa: string
    tipoEmpresa: string
    representanteEmpresa: string
    dniRepresentante: string
    direccionEmpresa: string
    distritoEmpresa: string
    provinciaEmpresa: string
    departamentoEmpresa: string

    nombreEstudiante: string
    dniEstudiante: string
    direccionEstudiante: string

    duracionPracticas: string

    representanteCargo: string

    gradoDecano: string
    nombreDecano: string
    dniDecano: string
    facultad: string
}